fix: convert numeric types when needed

+++ test-clients/src/main/java/com/hedera/services/bdd/suites/contract/hapi/ContractCallSuite.java
@@ -1470,7 +1470,7 @@ public class ContractCallSuite extends HapiApiSuite {
 							var receiverAccountId = accountIdFromHexedMirrorAddress(receiverAddr);
 							var receiver = asHeadlongAddress(asAddress(receiverAccountId));
 							final var transferCall = contractCall(TRANSFERRING_CONTRACT, "transferToAddress",
-									receiver, BigInteger.valueOf(10)
+									receiver, 10
 							)
 									.payingWith(ACCOUNT)
 									.logged();
